# app/docs/openapi_v0_6_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolymarketSyncRequest": {
        "additionalProperties": false,
        "properties": {
          "include_upcoming": {
            "default": false,
            "title": "Include Upcoming",
            "type": "boolean"
          },
          "max_finished": {
            "default": 2,
            "minimum": 0.0,
            "title": "Max Finished",
            "type": "integer"
          },
          "max_live": {
            "default": 2,
            "minimum": 0.0,
            "title": "Max Live",
            "type": "integer"
          },
          "max_upcoming": {
            "default": 2,
            "minimum": 0.0,
            "title": "Max Upcoming",
            "type": "integer"
          },
          "missing_only": {
            "default": false,
            "title": "Missing Only",
            "type": "boolean"
          },
          "probe_set": {
            "default": "today_nba",
            "enum": [
              "today_nba",
              "extras",
              "combined"
            ],
            "title": "Probe Set",
            "type": "string"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "title": "Steps",
            "type": "array"
          },
          "stream_max_outcomes": {
            "default": 30,
            "minimum": 1.0,
            "title": "Stream Max Outcomes",
            "type": "integer"
          },
          "stream_sample_count": {
            "default": 3,
            "minimum": 0.0,
            "title": "Stream Sample Count",
            "type": "integer"
          },
          "stream_sample_interval_sec": {
            "default": 1.0,
            "minimum": 0.0,
            "title": "Stream Sample Interval Sec",
            "type": "number"
          }
        },
        "title": "PolymarketSyncRequest",
        "type": "object"
      },
...
      "SyncTriggerResponse": {
        "additionalProperties": false,
        "properties": {
          "job_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Run Id"
          },
          "rows_read": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Read"
          },
          "rows_written": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Written"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "sync_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Run Id"
          }
        },
        "required": [
          "status"
        ],
        "title": "SyncTriggerResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "FastAPI service layer for Janus Cortex data platform",
    "title": "Janus Cortex API",
    "version": "0.7.6"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/events/{event_id}/markets": {
      "get": {
        "operationId": "get_event_markets_v1_events__event_id__markets_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Event Markets V1 Events  Event Id  Markets Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Event Markets",
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/events/{event_id}/odds/latest": {
      "get": {
        "operationId": "get_event_latest_odds_v1_events__event_id__odds_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "priced_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Priced Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Event Latest Odds V1 Events  Event Id  Odds Latest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Event Latest Odds",
        "tags": [
          "market-data"
        ]
      }
    },
...
    "/v1/markets/{market_id}": {
      "get": {
        "operationId": "get_market_v1_markets__market_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "market_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Market Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Market V1 Markets  Market Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Market",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/v1/markets/{market_id}/state/latest": {
      "get": {
        "operationId": "get_market_state_latest_v1_markets__market_id__state_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "market_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Market Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Market State Latest V1 Markets  Market Id  State Latest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Market State Latest",
        "tags": [
          "market-data"
        ]
      }
    },
...
    "/v1/outcomes/{outcome_id}/orderbook/history": {
      "get": {
        "operationId": "get_outcome_orderbook_history_v1_outcomes__outcome_id__orderbook_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "outcome_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Outcome Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time"
            }
          },
          {
            "in": "query",
            "name": "end_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_levels",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Levels",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "levels_per_side",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Levels Per Side",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Outcome Orderbook History V1 Outcomes  Outcome Id  Orderbook History Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Outcome Orderbook History",
        "tags": [
          "market-data"
        ]
      }
    },
    "/v1/outcomes/{outcome_id}/orderbook/latest": {
      "get": {
        "operationId": "get_outcome_orderbook_latest_v1_outcomes__outcome_id__orderbook_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "outcome_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Outcome Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "levels_per_side",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Levels Per Side",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Outcome Orderbook Latest V1 Outcomes  Outcome Id  Orderbook Latest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Outcome Orderbook Latest",
        "tags": [
          "market-data"
        ]
      }
    },
    "/v1/outcomes/{outcome_id}/prices/candles": {
      "get": {
        "operationId": "get_outcome_price_candles_v1_outcomes__outcome_id__prices_candles_get",
        "parameters": [
          {
            "in": "path",
            "name": "outcome_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Outcome Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timeframe",
            "required": false,
            "schema": {
              "default": "1m",
              "title": "Timeframe",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time"
            }
          },
          {
            "in": "query",
            "name": "end_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "asc",
            "required": false,
            "schema": {
              "default": false,
              "title": "Asc",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Outcome Price Candles V1 Outcomes  Outcome Id  Prices Candles Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Outcome Price Candles",
        "tags": [
          "market-data"
        ]
      }
    },
    "/v1/outcomes/{outcome_id}/prices/ticks": {
      "get": {
        "operationId": "get_outcome_price_ticks_v1_outcomes__outcome_id__prices_ticks_get",
        "parameters": [
          {
            "in": "path",
            "name": "outcome_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Outcome Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time"
            }
          },
          {
            "in": "query",
            "name": "end_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "asc",
            "required": false,
            "schema": {
              "default": false,
              "title": "Asc",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Outcome Price Ticks V1 Outcomes  Outcome Id  Prices Ticks Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Outcome Price Ticks",
        "tags": [
          "market-data"
        ]
      }
    },
...
    "/v1/sync/polymarket/markets": {
      "post": {
        "operationId": "sync_polymarket_markets_v1_sync_polymarket_markets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolymarketSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncTriggerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Polymarket Markets",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}